# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Snaptic API",
    "version": "1.0.0",
    "description": "API documentation for the Snaptic Attendance System.",
    "contact": {
      "name": "Tushar Sahu",
      "email": "sahutushar532@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in a cookie"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/attendance/start/{classId}": {
      "post": {
        "summary": "Start an attendance session for a class",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session started"
          }
        }
      }
    },
    "/attendance/mark": {
      "put": {
        "summary": "Mark attendance for a student manually",
        "tags": [
          "Attendance"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId",
                  "studentId",
                  "status"
                ],
                "properties": {
                  "sessionId": {
                    "type": "string"
                  },
                  "studentId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "present",
                      "absent"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance marked"
          }
        }
      }
    },
    "/attendance/submit/{sessionId}": {
      "post": {
        "summary": "Submit an attendance session for review",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session submitted"
          }
        }
      }
    },
    "/attendance/session/{sessionId}/reset": {
      "delete": {
        "summary": "Reset an attendance session",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session reset"
          }
        }
      }
    },
...
    "/attendance/today/{classId}": {
      "get": {
        "summary": "Get today's active session for a specific class",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Today's session for the class"
          }
        }
      }
    },
    "/attendance/session/{sessionId}/records": {
      "get": {
        "summary": "Get all attendance records for a session",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of attendance records"
          }
        }
      }
    },
...
  },
...
}